+++ R/dimreduc.R
@@ -1031,6 +1031,9 @@ setMethod(
 setValidity(
   Class = 'DimReduc',
   method = function(object) {
+    if (.GetSeuratCompat() < '5.0.0') {
+      return(TRUE)
+    }
     if (isFALSE(x = getOption(x = "Seurat.object.validate", default = TRUE))) {
       warn(
         message = paste("Not validating", class(x = object)[1L], "objects"),