+++ R/seurat.R
@@ -360,8 +360,11 @@ LoadSeuratRds <- function(file, ...) {
   object <- readRDS(file = file, ...)
   cache <- Tool(object = object, slot = 'SaveSeuratRds')
   reqd.cols <- c('layer', 'path', 'class', 'pkg', 'fxn', 'assay')
-  strict <- isTRUE(x = getOption(x = 'Seurat.io.rds.strict', default = FALSE))
-  emit <- ifelse(test = strict, yes = abort, no = warn)
+  emit <- ifelse(
+    test = isTRUE(x = getOption(x = 'Seurat.io.rds.strict', default = FALSE)),
+    yes = abort,
+    no = warn
+  )
   if (!is.null(x = cache)) {
     if (interactive()) {
       check_installed(pkg = 'fs', reason = 'for finding file paths')
@@ -385,7 +388,19 @@ LoadSeuratRds <- function(file, ...) {
       return(object)
     }
     # Check the files
-    exists <- fs::is_file(path = cache$path)
+    exists <- vapply(
+      X = cache$path,
+      FUN = function(x) {
+        x <- unlist(x = strsplit(x = x, split = ','))
+        res <- vector(mode = 'logical', length = length(x = x))
+        for (i in seq_along(along.with = x)) {
+          res[i] <- fs::is_file(path = x[i]) || fs::dir_exists(path = x[i])
+        }
+        return(all(res))
+      },
+      FUN.VALUE = logical(length = 1L),
+      USE.NAMES = FALSE
+    )
     exists[is.na(exists)] <- FALSE
     cache <- cache[exists, , drop = FALSE]
     if (!nrow(x = cache)) {
@@ -591,20 +606,6 @@ RenameAssays <- function(
   return(object)
 }
 
-saveRDS.Seurat <- function(
-  object = object,
-  file = NULL,
-  destdir = NULL,
-  ...
-) {
-  if (!is.null(x = destdir)) {
-    SaveSeuratBP(object, filename = basename(file), destdir = destdir, ...)
-  } else {
-    base::saveRDS(object = object, file = file, ...)
-    return(invisible(x = file))
-  }
-}
-
 #' Save and Load \code{Seurat} Objects from Rds files
 #'
 #' @param object A \code{\link{Seurat}} object
@@ -629,24 +630,55 @@ saveRDS.Seurat <- function(
 #' @order 1
 #'
 #' @examples
-#' if (requireNamespace("HDF5Array", quietly = TRUE) && requireNamespace("fs", quietly = TRUE)) {
-#'   pbmc_small[["disk"]] <- CreateAssay5Object(list(
-#'     mem = LayerData(pbmc_small, "counts"),
-#'     disk = as(LayerData(pbmc_small, "counts"), "HDF5Array")
-#'   ))
+#' if (requireNamespace("fs", quietly = TRUE)) {
+#'   # Write out with DelayedArray
+#'   if (requireNamespace("HDF5Array", quietly = TRUE)) {
+#'     pbmc <- pbmc_small
 #'
-#'   # Save `pbmc_small` to an Rds file
-#'   out <- tempfile(fileext = ".Rds")
-#'   SaveSeuratRds(pbmc_small, file = out)
+#'     pbmc[["disk"]] <- CreateAssay5Object(list(
+#'       mem = LayerData(pbmc, "counts"),
+#'       disk = as(LayerData(pbmc, "counts"), "HDF5Array")
+#'     ))
 #'
-#'   # Object cache
-#'   obj <- readRDS(out)
-#'   Tool(obj, "SaveSeuratRds")
+#'     # Save `pbmc` to an Rds file
+#'     out <- tempfile(fileext = ".Rds")
+#'     SaveSeuratRds(pbmc, file = out)
 #'
-#'   # Load the saved object with on-disk layers back into memory
-#'   pbmc2 <- LoadSeuratRds(out)
-#'   pbmc2
-#'   pbmc2[["disk"]]
+#'     # Object cache
+#'     obj <- readRDS(out)
+#'     Tool(obj, "SaveSeuratRds")
+#'
+#'     # Load the saved object with on-disk layers back into memory
+#'     pbmc2 <- LoadSeuratRds(out)
+#'     pbmc2
+#'     pbmc2[["disk"]]
+#'   }
+#'
+#'   # Write out with BPCells
+#'   if (requireNamespace("BPCells", quietly = TRUE)) {
+#'     pbmc <- pbmc_small
+#'
+#'     bpm <- BPCells::write_matrix_dir(LayerData(pbmc, "counts"), dir = tempfile())
+#'     bph <- BPCells::write_matrix_hdf5(
+#'       LayerData(pbmc, "counts"),
+#'       path = tempfile(fileext = ".h5"),
+#'       group = "counts"
+#'     )
+#'     pbmc[["disk"]] <- CreateAssay5Object(list(dir = bpm, h5 = bph))
+#'
+#'     # Save `pbmc` to an Rds file
+#'     out <- tempfile(fileext = ".Rds")
+#'     SaveSeuratRds(pbmc, file = out)
+#'
+#'     # Object cache
+#'     obj <- readRDS(out)
+#'     Tool(obj, "SaveSeuratRds")
+#'
+#'     # Load the saved object with on-disk layers back into memory
+#'     pbmc2 <- LoadSeuratRds(out)
+#'     pbmc2
+#'     pbmc2[["disk"]]
+#'   }
 #' }
 #'
 SaveSeuratRds <- function(
@@ -657,7 +689,7 @@ SaveSeuratRds <- function(
   ...
 ) {
   file <- file %||% file.path(getwd(), paste0(Project(object = object), '.Rds'))
-  file <- normalizePath(path = file, mustWork = FALSE)
+  file <- normalizePath(path = file, winslash = '/', mustWork = FALSE)
   # Cache v5 assays
   assays <- .FilterObjects(object = object, classes.keep = 'StdAssay')
   p <- progressor(along = assays, auto_finish = TRUE)
@@ -673,7 +705,7 @@ SaveSeuratRds <- function(
   )
   cache <- vector(mode = 'list', length = length(x = assays))
   names(x = cache) <- assays
-  tdir <- normalizePath(path = tempdir()) # because macOS is weird
+  tdir <- normalizePath(path = tempdir(), winslash = '/') # because macOS is weird
   destdir <- destdir %||% dirname(path = file)
   if (!is_na(x = destdir) || isTRUE(x = relative)) {
     check_installed(
@@ -692,6 +724,10 @@ SaveSeuratRds <- function(
       FUN = function(lyr) {
         ldat <- LayerData(object = object[[assay]], layer = lyr)
         path <- .FilePath(x = ldat)
+        path <- Filter(f = nzchar, x = path)
+        if (!length(x = path)) {
+          path <- NULL
+        }
         if (is.null(x = path)) {
           return(NULL)
         }
@@ -725,10 +761,6 @@ SaveSeuratRds <- function(
             class = 'sticky',
             amount = 0
           )
-          # df[i, 'path'] <- as.character(x = fs::file_move(
-          #   path = pth,
-          #   new_path = destdir
-          # ))
           df[i, 'path'] <- as.character(x = .FileMove(
             path = pth,
             new_path = destdir
@@ -786,175 +818,6 @@ SaveSeuratRds <- function(
   return(invisible(x = file))
 }
 
-#' Save \code{Seurat} Objects as RDS files with BPCells directories in same folder
-#'
-#' @param object A \code{\link{Seurat}} object
-#' @param file file name for \code{\link{Seurat}} object. defaults to
-#' \code{(paste0(Project(object), ".Rds"))}
-#' @param destdir Destination directory (to include both BPCells directory
-#' and \code{\link{Seurat}} object)
-#' @param relative Save relative paths instead of absolute ones. This is recommended
-#' if sharing the object folder
-#' @inheritDotParams base::saveRDS
-#'
-#' @return Invisibly returns \code{file}
-#'
-#' @export
-#'
-#' @template section-progressr
-#'
-#' @templateVar pkg fs
-#' @template note-reqdpkg
-#'
-#' @seealso \code{\link{saveRDS}()} \code{\link{readRDS}()}
-#'
-#' @order 1
-#'
-#' @examples
-#' if (requireNamespace("BPCells", quietly = TRUE) && requireNamespace("fs" , quietly = TRUE)) {
-#'   pbmc_small[["RNA5"]] <- as(object = pbmc_small[['RNA']], Class = 'Assay5')
-#'   pbmc_small[["RNA5"]]$counts <- BPCells::write_matrix_dir(
-#'   mat = pbmc_small[['RNA']]$counts,
-#'     dir = '~/pbmc_counts_BP/')
-#'
-#'   # Save `pbmc_small` to a folder with Rds file and BP Cells directory
-#'   SaveSeuratBP(pbmc_small,
-#'                filename = "pbmc_small.Rds",
-#'                destdir = "~/full_object/",
-#'                relative = TRUE)
-#'
-#'   # Load the saved object with on-disk layers back into memory
-#'   pbmc2 <- readRDS("~/full_object/pbmc_small.Rds")
-#'   pbmc2
-#'   pbmc2[["RNA5"]]$counts
-#' }
-#'
-SaveSeuratBP <- function(
-  object,
-  filename = NULL,
-  destdir = NULL,
-  relative = FALSE,
-  ...
-) {
-  filename <- filename %||% paste0(Project(object = object), '.Rds')
-  # Cache v5 assays
-  assays <- .FilterObjects(object = object, classes.keep = 'StdAssay')
-  p <- progressor(along = assays, auto_finish = TRUE)
-  on.exit(expr = p(type = 'finish'), add = TRUE)
-  p(
-    message = paste(
-      "Looking for BPCells Directories in",
-      length(x = assays),
-      "assays"
-    ),
-    class = 'sticky',
-    amount = 0
-  )
-  cache <- vector(mode = 'list', length = length(x = assays))
-  names(x = cache) <- assays
-  destdir <- destdir %||% dirname(path = file)
-  if (!is_na(x = destdir) || isTRUE(x = relative)) {
-    check_installed(
-      pkg = 'fs',
-      reason = 'for moving on-disk matrices'
-    )
-  }
-  moved.paths <- c()
-  for (assay in assays) {
-    p(
-      message = paste("Searching through assay", assay),
-      class = 'sticky',
-      amount = 0
-    )
-    df <- lapply(
-      X = Layers(object = object[[assay]]),
-      FUN = function(lyr) {
-        ldat <- LayerData(object = object[[assay]], layer = lyr)
-        path <- .FilePath(x = ldat)
-        if (is.null(x = path)) {
-          return(NULL)
-        }
-        return(data.frame(
-          layer = rep(lyr, length(path)),
-          path = path,
-          class = rep(paste(class(x = ldat), collapse = ','), length(path)),
-          pkg = rep(.ClassPkg(object = ldat), length(path))
-        ))
-      }
-    )
-    df <- do.call(what = 'rbind', args = df)
-    if (is.null(x = df) || !nrow(x = df)) {
-      p(message = "No on-disk layers found", class = 'sticky', amount = 0)
-      next
-    }
-    if (!is_na(x = destdir)) {
-      for (i in seq_len(length.out = nrow(x = df))) {
-        pth <- df$path[i]
-        p(
-          message = paste(
-            "Moving layer",
-            sQuote(x = df$layer[i]),
-            "to",
-            sQuote(x = destdir)
-          ),
-          class = 'sticky',
-          amount = 0
-        )
-        if (!pth %in% names(moved.paths)){
-          df[i, 'path'] <- tryCatch(
-            expr = as.character(x = .FileMove(
-              path = pth,
-              new_path = destdir
-            )), error = function(e) {
-              warning("Error: ", e, call. = FALSE)
-              # Return original path if can't locate new path
-              return(pth)
-            }
-          )
-          # Record that this dir was already moved in case a new layer has the same source
-          moved.paths[[pth]] <- df[i, 'path']
-        } else{
-          df[i, 'path'] <- moved.paths[pth]
-        }
-      }
-    }
-    if (isTRUE(x = relative)) {
-      p(
-        message = paste(
-          "Adjusting paths to be relative to",
-          sQuote(x = dirname(path = destdir), q = FALSE)
-        ),
-        class = 'sticky',
-        amount = 0
-      )
-      df$path <- as.character(x = fs::path_rel(
-        path = df$path,
-        start = dirname(path = destdir)
-      ))
-    }
-    df$assay <- assay
-    for (layer in unique(df$layer)) {
-      warning("Changing path in object to point to new BPCells directory location",
-              call. = FALSE, immediate. = TRUE)
-      ldat <- LayerData(object[[assay]], layer = layer)
-      matrices <- BPCells::all_matrix_inputs(ldat)
-      matrix.dirs <- which(sapply(matrices, function(x) inherits(x, "MatrixDir")))
-      if (length(matrix.dirs) == nrow(df[df$layer == layer, ])) {
-        for (i in 1:length(matrix.dirs)) {
-          matrices[[matrix.dirs[i]]]@dir <- df[df$layer == layer, ]$path[i]
-        }
-        BPCells::all_matrix_inputs(ldat) <- matrices
-        LayerData(object[[assay]], layer = layer) <- ldat
-      } else {
-        stop("Directories to replace is not equal to length of directories")
-      }
-    }
-    p()
-  }
-  saveRDS(object = object, file = file.path(destdir, filename), ...)
-  return(invisible(x = filename))
-}
-
 #' Update old Seurat object to accommodate new features
 #'
 #' Updates Seurat objects to new structure for storing data/calculations.
@@ -966,7 +829,6 @@ SaveSeuratBP <- function(
 #' @return Returns a Seurat object compatible with latest changes
 #'
 #' @importFrom methods .hasSlot new slot
-#' @importFrom utils packageVersion
 #'
 #' @export
 #'
@@ -1379,9 +1241,13 @@ CreateSeuratObject.default <- function(
   ...
 ) {
   assay.version <- getOption(x = 'Seurat.object.assay.version', default = 'v5')
-  if(!inherits(counts, what = "dgCMatrix") && assay.version == 'v3'){
-    message("Counts matrix provided is not sparse. Creating V5 assay in Seurat Object.")
-    assay.version = 'v5'
+  if (.GetSeuratCompat() < '5.0.0') {
+    assay.version <- 'v3'
+  } else if (!inherits(counts, what = "AnyMatrix") && assay.version == 'v3') {
+    message(
+      "Counts matrix provided is not sparse; vreating v5 assay in Seurat object"
+    )
+    assay.version <- 'v5'
   }
   assay.data <- if (tolower(x = assay.version) == 'v3') {
     assay.data <- CreateAssayObject(
@@ -1472,15 +1338,6 @@ CreateSeuratObject.Assay <- function(
   ))
   options(op)
   object[['orig.ident']] <- idents
-  # Add provided meta data
-  if (!is.null(x = meta.data)) {
-    tryCatch(
-      expr = object[[]] <- meta.data,
-      error = function(e) {
-        warning(e$message, call. = FALSE, immediate. = TRUE)
-      }
-    )
-  }
   # Calculate nCount and nFeature
   calcN_option <- getOption(
     x = 'Seurat.object.assay.calcn',
@@ -1493,6 +1350,15 @@ CreateSeuratObject.Assay <- function(
       names(x = ncalc) <- paste(names(x = ncalc), assay, sep = '_')
       object[[]] <- ncalc
     }
+  }
+  # Add provided meta data
+  if (!is.null(x = meta.data)) {
+    tryCatch(
+      expr = object[[]] <- meta.data,
+      error = function(e) {
+        warning(e$message, call. = FALSE, immediate. = TRUE)
+      }
+    )
   }
   # Validate and return
   validObject(object = object)
@@ -1647,7 +1513,6 @@ Features.Seurat <- function(x, assay = NULL, ...) {
   assay <- match.arg(arg = assay, choices = Assays(object = x))
   return(Features(x = x[[assay]], ...))
 }
-
 
 #' @param vars List of all variables to fetch, use keyword \dQuote{ident} to
 #' pull identity classes
@@ -4317,6 +4182,9 @@ setMethod(
     value = 'Assay'
   ),
   definition = function(x, i, ..., value) {
+    if (.GetSeuratCompat() < '5.0.0') {
+      return(callNextMethod(x = x, i = i, value = value))
+    }
     validObject(object = value)
     i <- make.names(names = i)
     # Checks for if the assay or name already exists
@@ -5587,6 +5455,9 @@ setMethod(
 setValidity(
   Class = 'Seurat',
   method = function(object) {
+    if (.GetSeuratCompat() < '5.0.0') {
+      return(TRUE)
+    }
     if (isFALSE(x = getOption(x = "Seurat.object.validate", default = TRUE))) {
       warn(
         message = paste("Not validating", class(x = object)[1L], "objects"),