+++ R/zzz.R
@@ -1,8 +1,9 @@
 #' @importFrom sp bbox over
 #' @importFrom Rcpp evalCpp
+#' @importFrom Matrix nnzero
 #' @importFrom progressr progressor
-#' @importFrom utils head tail upgrade
 #' @importFrom lifecycle deprecated is_present
+#' @importFrom utils head packageVersion tail upgrade
 #' @importFrom methods new setClass setClassUnion setGeneric setMethod
 #' setOldClass setValidity show slot slot<- validObject
 #' @importFrom rlang abort arg_match arg_match0 caller_env check_installed
@@ -19,12 +20,6 @@ NULL
 #' @rdname SeuratObject-package
 #'
 "_PACKAGE"
-
-#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-# Package options
-#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-
-default.options <- list()
 
 #%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 # Options
@@ -83,11 +78,21 @@ Seurat.options <- list(
   Seurat.coords.short_range = 'max',
   Seurat.input.sparse_ratio = 0.4,
   Seurat.io.rds.strict = FALSE,
+  Seurat.object.assay.brackets = 'v5',
   Seurat.object.assay.calcn = NULL,
   Seurat.object.assay.version = 'v5',
   Seurat.object.assay.v3.missing_layer = 'matrix',
   Seurat.object.project = 'SeuratProject',
   progressr.clear = FALSE
+)
+
+#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+# Built With
+#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+.BuiltWith <- c(
+  R = format(x = getRversion()),
+  Matrix = format(x = packageVersion(pkg = "Matrix"))
 )
 
 #%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
@@ -243,8 +248,6 @@ setOldClass(Classes = 'package_version')
 #'
 #' @return \code{TRUE} if \pkg{SeuratObject} and/or \pkg{Seurat}
 #'
-#' @importFrom utils packageVersion
-#'
 #' @keywords internal
 #'
 #' @export
@@ -376,6 +379,62 @@ setOldClass(Classes = 'package_version')
   return(idx)
 }
 
+#' Round Version Information
+#'
+#' @param current A package version
+#'
+#' @return ...
+#'
+#' @keywords internal
+#'
+#' @noRd
+#'
+.RoundVersion <- function(current) {
+  current <- as.character(x = numeric_version(x = current, strict = TRUE))
+  current <- unlist(x = strsplit(x = current, split = '\\.'))
+  if (length(x = current) > 4L) {
+    if (length(x = current) > 4L) {
+      current[4L] <- paste(
+        current[seq.int(from = 4L, to = length(x = current))],
+        collapse = '.'
+      )
+      current <- current[1:4]
+    }
+  }
+  names(x = current) <- c('major', 'minor', 'patch', 'devel')[seq_along(along.with = current)]
+  if (!is_na(x = current['devel'])) {
+    if (all(current[c('minor', 'patch')] == '9')) {
+      current['major'] <- as.character(x = as.integer(x = current['major']) + 1L)
+      current[c('minor', 'patch')] <- '0'
+    } else if (current['patch'] == '0') {
+      current['minor'] <- as.character(x = as.integer(x = current['minor']) + 1L)
+      current['patch'] <- '0'
+    } else {
+      current['patch'] <- as.character(x = as.integer(x = current['patch']) + 1L)
+    }
+    current <- current[c('major', 'minor', 'patch')]
+  }
+  current <- vapply(
+    X = current,
+    FUN = as.integer,
+    FUN.VALUE = integer(length = 1L),
+    USE.NAMES = TRUE
+  )
+  if (!is_na(x = current['devel'])) {
+    if (all(current[c('minor', 'patch')] == '9')) {
+      current['major'] <- as.character(x = as.integer(x = current['major']) + 1L)
+      current[c('minor', 'patch')] <- '0'
+    } else if (current['patch'] == '0') {
+      current['minor'] <- as.character(x = as.integer(x = current['minor']) + 1L)
+      current['patch'] <- '0'
+    } else {
+      current['patch'] <- as.character(x = as.integer(x = current['patch']) + 1L)
+    }
+    current <- current[c('major', 'minor', 'patch')]
+  }
+  return(current)
+}
+
 #' Index of Names
 #'
 #' Get the index of row- or column-names
@@ -437,10 +496,44 @@ NameIndex <- function(x, names, MARGIN) {
 # Hooks
 #%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
+.onAttach <- function(libname, pkgname) {
+  for (i in names(x = .BuiltWith)) {
+    current <- switch(EXPR = i, R = getRversion(), packageVersion(pkg = i))
+    if (current > .BuiltWith[i]) {
+      msg <- paste(
+        sQuote(x = pkgname),
+        "was built",
+        switch(
+          EXPR = i,
+          R = "under R",
+          paste("with package", sQuote(x = i))
+        ),
+        .BuiltWith[i],
+        "but the current version is",
+        paste0(current, ';'),
+        "it is recomended that you reinstall ",
+        sQuote(x = pkgname),
+        " as the ABI for",
+        switch(EXPR = i, R = i, sQuote(x = i)),
+        "may have changed"
+      )
+      packageStartupMessage(paste(strwrap(x = msg), collapse = '\n'))
+    }
+  }
+}
+
 .onLoad <- function(libname, pkgname) {
   toset <- setdiff(x = names(x = Seurat.options), y = names(x = options()))
   if (length(x = toset)) {
     options(Seurat.options[toset])
   }
+  setHook(
+    hookName = packageEvent(pkgname = 'Seurat', event = 'onLoad'),
+    value = .SetSeuratCompat
+  )
+  setHook(
+    hookName = packageEvent(pkgname = 'Seurat', event = 'attach'),
+    value = .SeuratCompatMessage
+  )
   return(invisible(x = NULL))
 }